update LCD and TCN
remove the LCD function from the "TCN" file, as they are in the "LCD" file

+++ laba3/atmel/laba3/laba3/TCN75.h
@@ -10,6 +10,7 @@
 #define TCN75_H_
 
 #include <avr/io.h>
+#include "LCD.h"
 
 typedef enum _Sequence
 {
@@ -19,6 +20,6 @@ typedef enum _Sequence
 	TWI_Sequence_Stop
 }_Sequence;
 
-uint16_t TCN75ReadTemp(uint16_t adres);
+uint16_t TCN75ReadTemp(uint8_t adres);
 
 #endif
